BUDGO-12: jackson LocalDate serialization config
