+++ src/main/java/com/godev/budgetgo/api/rest/operation/dto/OperationInfoDto.java
@@ -1,9 +1,7 @@
 package com.godev.budgetgo.api.rest.operation.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.godev.budgetgo.api.rest.user.dto.UserInfoDto;
-import com.godev.budgetgo.infra.json.LocalDateSerializer;
 import io.swagger.annotations.ApiModel;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -22,15 +20,12 @@ public class OperationInfoDto {
     @JsonProperty("category")
     private CategoryInfoDto categoryInfoDto;
 
-    @JsonSerialize(using = LocalDateSerializer.class)
     private LocalDate date;
 
     private String description;
 
-    @JsonSerialize(using = LocalDateSerializer.class)
     private LocalDate dateCreated;
 
-    @JsonSerialize(using = LocalDateSerializer.class)
     private LocalDate dateModified;
 
     @JsonProperty("lastEditor")
