+++ src/test/java/com/godev/budgetgo/infra/jackson/LocalDateDeserializerTest.java
@@ -1,4 +1,4 @@
-package com.godev.budgetgo.infra.json;
+package com.godev.budgetgo.infra.jackson;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
